split out ResourceName component, fix line-breaks in flavor data
