+++ plugins/resources/app/javascript/app/components/project/resource.jsx
@@ -1,22 +1,6 @@
-import { OverlayTrigger, Tooltip } from 'react-bootstrap';
-
 import { Unit } from '../../unit';
 import { byUIString, t } from '../../utils';
-
-const ResourceName = ({name, flavorData}) => {
-  if (!flavorData.primary || !flavorData.secondary) {
-    return <div className='col-md-2 text-right'>{name}</div>;
-  }
-
-  let tooltip = <Tooltip id={`tooltip-${name}`} className='tooltip-no-break'>{flavorData.secondary}</Tooltip>;
-  // TODO: line breaks make the .small.text-muted.flavor-data look ugly -- maybe place below bar like we had for .resource-error?
-  return (
-    <div className='col-md-2 text-right'>
-      <OverlayTrigger overlay={tooltip} placement='right' delayShow={300} delayHide={150}><span>{name}</span></OverlayTrigger>
-      <div className='small text-muted flavor-data'>{flavorData.primary}</div>
-    </div>
-  );
-};
+import ResourceName from '../../components/resource_name';
 
 const valueWithUnit = (value, unit) => {
   const title = unit.name !== '' ? `${value} ${unit.name}` : undefined;
